requirements.txt
requests
gitpython
docker
tqdm
passlib
kubernetes